% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assist_networks.R
\name{assist_net}
\alias{assist_net}
\title{Assist Network}
\usage{
assist_net(team, node_col, season, rmv_bench, tree, three_weights,
  message = NA)
}
\arguments{
\item{team}{Team to create network for}

\item{node_col}{Color of nodes in network}

\item{season}{Season, as a character,  (e.g. "2018-19"), or vector of ESPN gameIDS
for which data to use in network. Currently only handles the current season worth of data.}

\item{rmv_bench}{Logical indicating whether to remove players who do not factor in the network.}

\item{tree}{Logicial indicating whether to draw the network as a tree (recommended for single games).
If FALSE, the network will be drawn as a circle (recommended for entire seasons).}

\item{three_weights}{Logical indicating whether to give extra weight for assisted three point shots.
If TRUE, assisted three-point shots will be given weight 1.5 (as opposed to weight 1).}

\item{message}{User supplied plot title to overwrite default plot title, if desired. Default = NA.}
}
\description{
Renders assist network for game or entire season
}
