% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winprob.R
\name{average_score_diff}
\alias{average_score_diff}
\title{Average Score Differential}
\usage{
average_score_diff(game_id)
}
\arguments{
\item{game_id}{ESPN game_id for which to compute average score differential.}
}
\value{
Average score differential
}
\description{
Computes Time-Based Average Score Differential for the Home Team
}
