% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_win_prob.R
\name{average_win_prob}
\alias{average_win_prob}
\title{Average Win Probability}
\usage{
average_win_prob(game_id, include_spread = T)
}
\arguments{
\item{game_id}{ESPN game_id for which to compute average win probability}

\item{include_spread}{Logical, whether to include pre-game spread in Win Probability
calculations. Default = True.}
}
\value{
Average Win Probability
}
\description{
Computes Time-Based Average Win Probability for the Home Team
}
