% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assist_networks.R
\name{circle_assist_net}
\alias{circle_assist_net}
\title{Circle Assist Network}
\usage{
circle_assist_net(team, season, highlight_player = NA,
  highlight_color = NA, three_weights = T, threshold = 0,
  message = NA)
}
\arguments{
\item{team}{Team to create network for}

\item{season}{Season, as a character,  (e.g. "2018-19"), or vector of ESPN game_ids.
for which data to use in network. Currently only handles the current season worth of data.}

\item{highlight_player}{Name of player to highlight in assist network. `NA` yields full team assist
network with no player highlighting. Default = `NA`.}

\item{highlight_color}{Color of player links to be highlighted. `NA` if ```highlight_player``` is `NA`.}

\item{three_weights}{Logical indicating whether to give extra weight for assisted three point shots.
If TRUE, assisted three-point shots will be given weight 1.5 (as opposed to weight 1). Default = `TRUE`.}

\item{threshold}{Number between 0-1 indicating minimum percentage of team assists/baskets a player needs to exceed to be included in network. Default = 0.}

\item{message}{User supplied plot title to overwrite default plot title, if desired. Default = `NA`.}
}
\value{
List with network statistics
\itemize{
 \item{"clust_coeff"} - Network Clustering Coefficient
 \item{"page_ranks"} - Player Page Ranks in network
 \item{"hub_scores"} - Player Hub Scores in network
 \item{"auth_scores"} - Player Authority Scores in network
 \item{"ast_freq"} - Player percentage of team's assists
 \item{"shot_freq"} - Player percenatge of scoring on team's assisted baskets
 }
}
\description{
Renders circlized assist network for game or entire season and allows for highlight
of a single player if desired.
}
