% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{court}
\alias{court}
\title{A data frame with NCAA court dimensions.}
\format{A data frame with NCAA court dimensions}
\usage{
court
}
\description{
A data frame with NCAA court dimensions.
}
\details{
use this dataframe with geom_polygon to draw a NCAA court.
}
\examples{
data(court)

}
\keyword{datasets}
