% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dict}
\alias{dict}
\title{A data frame with team names from various college basketball sites}
\format{A data frame with team names from various college basketball sites
\describe{
 \item{NCAA}{the name of the team, as listed on the NCAA website}
 \item{ESPN}{the name of the team, as listed on the ESPN URLs}
 \item{ESPN_PBP}{the name of the team, as listed on the ESPN Play-By-Play logs}
 \item{Warren_Nolan}{the name of the team, as listed on WarrenNolan.com}
 \item{Trank}{the name of the team, as listed on barttorvik.com}
 \item{name_247}{the name of the team, as listed on 247Sports.com}
}}
\usage{
dict
}
\description{
A data frame with team names from various college basketball sites
}
\details{
Convert between team names in various systems
}
\examples{
data(dict)

}
\keyword{datasets}
