% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winprob.R
\name{game_flow}
\alias{game_flow}
\title{Game Flow Chart}
\usage{
game_flow(game_id, home_col, away_col)
}
\arguments{
\item{game_id}{ESPN game_id for which to render chart}

\item{home_col}{Color of home team for chart}

\item{away_col}{Color of away team for chart}
}
\description{
Renders Game Flow Chart
}
