% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_boxscore.R
\name{get_boxscore}
\alias{get_boxscore}
\title{Get Boxscore}
\usage{
get_boxscore(game_id)
}
\arguments{
\item{game_id}{ESPN game_id for which to scrape boxscore.}
}
\value{
A named list containing two dataframes with box score for each team.
  First team in list is away team, second is home team.
}
\description{
Gets boxscores for each team for a given game.
}
