% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncaa_pbp_scraper.R
\name{get_game_IDs}
\alias{get_game_IDs}
\title{Get Team GameIDs}
\usage{
get_game_IDs(team)
}
\arguments{
\item{team}{Team to get gameIDs}
}
\value{
A vector of the team's ESPN gameIDs for current season
}
\description{
Gets team gameIDs for current season.
}
