% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shot_charts.R
\name{get_game_plot}
\alias{get_game_plot}
\title{Plot all shots from one game}
\usage{
get_game_plot(game_id, heatmap = F)
}
\arguments{
\item{game_id}{ESPN game_id}

\item{heatmap}{Use a density type heatmap}
}
\description{
Plots game shot locations from one game.
}
