% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_master_schedule.R
\name{get_master_schedule}
\alias{get_master_schedule}
\title{Get Master Schedule}
\usage{
get_master_schedule(date)
}
\arguments{
\item{date}{Date for which to get schedule (YYYY-MM-DD)}
}
\value{
A data-frame of the day's schedule of games
}
\description{
Gets schedule for all games on a given date
}
