% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pbp.R
\name{get_pbp}
\alias{get_pbp}
\title{Get Team Play-by-Play Data}
\usage{
get_pbp(team, season = current_season)
}
\arguments{
\item{team}{Team to get Play-by-Play data for}

\item{season}{Season to get schedule for in form "2019-20". Default equals current season.}
}
\value{
A data-frame of the team's Play-by-Play data for the specified season.
}
\description{
Scrapes the current season's Play-by-Play data for desired team. Team
is assumed to be the ESPN team name, which can be looked up in the ids
dataframe.
}
