% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncaa_pbp_scraper.R
\name{get_pbp_game}
\alias{get_pbp_game}
\title{Get Game Play-by-Play Data}
\usage{
get_pbp_game(gameIDs)
}
\arguments{
\item{gameIDs}{Vector of ESPN game-IDs}
}
\value{
A data-frame of the Play-by-Play data fror desired games.
}
\description{
Scrapes ESPN Play-by-Play data for the desired games.
}
