% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_roster.R
\name{get_roster}
\alias{get_roster}
\title{Get Team Roster}
\usage{
get_roster(team, season = current_season)
}
\arguments{
\item{team}{Team to get roster for}

\item{season}{Season to get roster for. In form "2019-20". Default equals current season.}
}
\value{
A data frame of the team's roster for requested season
}
\description{
Gets team Roster for requested season.
}
