% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_schedule.R
\name{get_schedule}
\alias{get_schedule}
\title{Get Team Schedule}
\usage{
get_schedule(team, season = current_season)
}
\arguments{
\item{team}{Team to get schedule for}

\item{season}{Season to get schedule for. In form "2019-20". Default equals current season.}
}
\value{
A data frame of the team's schedule for requested season
}
\description{
Gets team schedule for requested season.
}
