% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncaa_get_shot_locs.R
\name{get_shot_locs}
\alias{get_shot_locs}
\title{Get Shot Locations}
\usage{
get_shot_locs(gameId)
}
\arguments{
\item{game_ids}{Vector of ESPN game-IDs}
}
\value{
A data-frame with shot details (including location) for given game_ids
}
\description{
Extracts Shot location data for a specific game
}
\examples{
get_shot_locs(401083556)
}
