% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncaa_get_team_plot.R
\name{get_team_plot}
\alias{get_team_plot}
\title{Plot all shots for a team}
\usage{
get_team_plot(df, team, heatmap = F)
}
\arguments{
\item{df}{dataframe of shot locations}

\item{team}{Team Name}

\item{heatmap}{Use a density type heatmap}
}
\description{
Plots team shot locations from one game or multiple games. DF can be obtained from
get_shot_locs
}
