% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/season_boxscore.R
\name{season_boxscore}
\alias{season_boxscore}
\title{Get Season Box Score}
\usage{
season_boxscore(team, season = current_season, aggregate = "average")
}
\arguments{
\item{team}{Team to get schedule for}

\item{season}{Season to get schedule for. In form "2020-21". Default equals current season.}

\item{aggregate}{One of 'average' (per-game average statistics), 'total' (sums of season stats) or 'raw' 
(just return all box scores binded together). 'average' is the default.}
}
\value{
A data frame of the team's schedule for requested season
}
\description{
Gets team schedule for requested season.
}
