% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shot_charts.R
\name{team_shot_chart}
\alias{team_shot_chart}
\title{Plot all shots for a team}
\usage{
team_shot_chart(game_ids, team, heatmap = F)
}
\arguments{
\item{game_ids}{Vector of ESPN game_ids}

\item{team}{Team Name}

\item{heatmap}{Use a density type heatmap (Default = FALSE)}
}
\description{
Plots team shot locations from one game or multiple games.
}
