% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winprob.R
\name{wp_chart}
\alias{wp_chart}
\title{Win Probability Chart}
\usage{
wp_chart(gameID, home_col, away_col, show_legend = T)
}
\arguments{
\item{gameID}{ESPN gameID for which to create chart}

\item{home_col}{Color of home team for chart}

\item{show_legend}{Logical indicating whether or not to display legend and min win probability
on the chart. Default = TRUE}

\item{home_col}{Color of away team for chart}
}
\description{
Renders win proability chart for desired game
}
