% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wp_chart_new.R
\name{wp_chart_new}
\alias{wp_chart_new}
\title{Win Probability Chart}
\usage{
wp_chart_new(game_id, include_spread = T, show_labels = T)
}
\arguments{
\item{game_id}{ESPN game_id for which to render chart}

\item{include_spread}{Logical, whether to include pre-game spread in Win Probability
calculations. Default = True.}

\item{show_labels}{Logical whether Game Exictement Index and Minimum
Win Probability metrics should be displayed on the plot. Default = TRUE.}
}
\description{
Renders Win Probability Charts in ggplot
}
