% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assist_networks.R
\name{assist_net}
\alias{assist_net}
\title{Assist Network}
\usage{
assist_net(team, season, node_col, three_weights = T, threshold = 0,
  message = NA)
}
\arguments{
\item{team}{Team to create network for}

\item{season}{Season, as a character,  (e.g. "2018-19"), or vector of ESPN game_ids.
for which data to use in network. Currently only handles the current season worth of data.}

\item{node_col}{Color of nodes in network}

\item{three_weights}{Logical indicating whether to give extra weight for assisted three point shots.
If TRUE, assisted three-point shots will be given weight 1.5 (as opposed to weight 1). Default = `TRUE`.}

\item{threshold}{Number between 0-1 indicating minimum percentage of team assists/baskets a player needs to exceed to be included in network. Default = 0.}

\item{message}{User supplied plot title to overwrite default plot title, if desired. Default = `NA`.}
}
\value{
List with network statistics
\itemize{
 \item{"clust_coeff"} - Network Clustering Coefficient
 \item{"page_ranks"} - Player Page Ranks in network
 \item{"hub_scores"} - Player Hub Scores in network
 \item{"auth_scores"} - Player Authority Scores in network
 \item{"ast_freq"} - Player percentage of team's assists
 \item{"shot_freq"} - Player percenatge of scoring on team's assisted baskets
 }
}
\description{
Renders assist network for game or entire season
}
