% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winprob.R
\name{average_win_prob}
\alias{average_win_prob}
\title{Average Win Probability}
\usage{
average_win_prob(game_id)
}
\arguments{
\item{game_id}{ESPN game_id for which to compute average win probability}
}
\value{
Average Win Probability
}
\description{
Computes Time-Based Average Win Probability for the Home Team
}
