% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_game_ids.R
\name{get_game_ids}
\alias{get_game_ids}
\title{Get Team game_ids}
\usage{
get_game_ids(team, season = current_season)
}
\arguments{
\item{team}{Team to get game_ids}

\item{season}{Season to get schedule for in form "2019-20". Default equals current season.}
}
\value{
A vector of the team's ESPN game_ids for current season
}
\description{
Gets team game_ids for current season.
}
