% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncaa_pbp_scraper.R
\name{get_game_ids}
\alias{get_game_ids}
\title{Get Team game_ids}
\usage{
get_game_ids(team)
}
\arguments{
\item{team}{Team to get game_ids}
}
\value{
A vector of the team's ESPN game_ids for current season
}
\description{
Gets team game_ids for current season.
}
