% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncaa_get_game_plot.R
\name{get_game_plot}
\alias{get_game_plot}
\title{Plot all shots from one game}
\usage{
get_game_plot(shot_loc_df, heatmap = F)
}
\arguments{
\item{shot_loc_df}{dataframe of shot locations}

\item{heatmap}{Use a density type heatmap}
}
\description{
Plots game shot locations from one game. DF can be obtained from
get_shot_locs
}
