% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncaa_pbp_scraper.R
\name{get_master_schedule}
\alias{get_master_schedule}
\title{Get Master Schedule}
\usage{
get_master_schedule(year, month, day)
}
\arguments{
\item{year}{year for which to get master schedule}

\item{month}{month for which to get master schedule}

\item{day}{day for which to get master schedule}
}
\value{
A data-frame of the day's schedule of games
}
\description{
Gets schedule for all games on a given date
}
