% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncaa_pbp_scraper.R
\name{get_pbp}
\alias{get_pbp}
\title{Get Team Play-by-Play Data}
\usage{
get_pbp(team)
}
\arguments{
\item{team}{Team to get Play-by-Play data for}
}
\value{
A data-frame of the team's Play-by-Play data for the current season
}
\description{
Scrapes the current season's Play-by-Play data for desired team. Team
is assumed to be the ESPN team name, which can be looked up in the ids
dataframe.
}
