% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pbp_game.R
\name{get_pbp_game}
\alias{get_pbp_game}
\title{Get Game Play-by-Play Data}
\usage{
get_pbp_game(game_ids, extra_parse = T)
}
\arguments{
\item{game_ids}{Vector of ESPN game-IDs}

\item{extra_parse}{Logical whether to link shot variables and possesion parsing
(Default = TRUE).}
}
\value{
A data-frame of the Play-by-Play data fror desired games.
}
\description{
Scrapes ESPN Play-by-Play data for the desired games.
}
