% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncaa_pbp_scraper.R
\name{get_pbp_game}
\alias{get_pbp_game}
\title{Get Game Play-by-Play Data}
\usage{
get_pbp_game(game_ids)
}
\arguments{
\item{game_ids}{Vector of ESPN game-IDs}

\item{win_prob}{Logical whether to return win probability from home team perspective
on each play of the game. Default = F.}
}
\value{
A data-frame of the Play-by-Play data fror desired games.
}
\description{
Scrapes ESPN Play-by-Play data for the desired games.
}
