% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncaa_pbp_scraper.R
\name{get_schedule}
\alias{get_schedule}
\title{Get Team Schedule}
\usage{
get_schedule(team)
}
\arguments{
\item{team}{Team to get schedule for}
}
\value{
A data-frame of the team's schedule for current season
}
\description{
Gets team schedule for current season.
}
