% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wp_chart.R
\name{wp_chart}
\alias{wp_chart}
\title{Win Probability Chart}
\usage{
wp_chart(game_id, home_col, away_col, include_spread = T, show_legend = T)
}
\arguments{
\item{game_id}{ESPN game_id for which to create chart}

\item{home_col}{Color of home team for chart}

\item{away_col}{Color of away team for chart}

\item{include_spread}{Logical, whether to include pre-game spread in Win Probability
calculations. Default = True.}

\item{show_legend}{Logical indicating whether or not to display legend and min win probability
on the chart. Default = TRUE}
}
\description{
Renders win proability chart for desired game
}
