% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wp_chart_new.R
\name{wp_chart_new}
\alias{wp_chart_new}
\title{Win Probability Chart}
\usage{
wp_chart_new(
  game_id,
  home_col = NULL,
  away_col = NULL,
  home_team = NULL,
  away_team = NULL,
  include_spread = T,
  show_labels = T
)
}
\arguments{
\item{game_id}{ESPN game_id for which to render chart}

\item{home_col}{Color of home team for chart. Defaults to primary color if NULL}

\item{away_col}{Color of away team for chart. Defaults to primary color if NULL}

\item{home_team}{Home team. Defaults to ESPN PBP Name}

\item{away_team}{Away team. Defaults to ESPN PBP Name}

\item{include_spread}{Logical, whether to include pre-game spread in Win Probability
calculations. Default = True.}

\item{show_labels}{Logical whether Game Exictement Index and Minimum
Win Probability metrics should be displayed on the plot. Default = TRUE.}
}
\description{
Renders Win Probability Charts in ggplot
}
